feat: enhance docker build workflow to support multi-platform images
This commit introduces changes to the Docker build workflow to
support building and publishing multi-platform images for both
release and development builds. It includes separate jobs for
building ARM64 and AMD64 images, creating manifest lists, and
pushing them to Docker Hub. Additionally, a new Dockerfile for
release builds is added, along with modifications to the existing
Dockerfile.